refactor(path_complete): split to own module
Split `main::FilePathCompleter` to own module and rename it as
`path_complete::PathComplete`. While at it, update SPDX platters to point
out GPL-2.0-or-later, so that it is clear that all the modules can be also
relicesed with future versions of GPL. Finally, document the modules, in
order to manage rising complexity.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>